+++ pubspec.yaml
@@ -10,7 +10,7 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.6
-  provider: ^6.1.2
+  flutter_riverpod: ^2.5.1
 
 dev_dependencies:
   flutter_test:
